Add a global varibalbe depending on the page.
I do this because I don't know any way to idetify what page or html tag I am useing. This way on the head of the html page the page that you are currently looking at will always be highlited. If you have any suggestion on how to idenify the page and/or html tag via javascript please tell me in the issues thanks.